pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "security-recon"
version = "0.1.0"
description = "Security reconciliation utilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"sqlalchemy>=2.0",
	"pymysql[rsa]>=1.1",
	"PyYAML>=6.0",
	"psycopg[binary]>=3.1",
	"streamlit",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src/main/python"]

[tool.pyright]
extraPaths = ["src/main/python"]

[tool.pytest.ini_options]
testpaths = ["src/test/python/security_recon_tests"]
